+++ packages/core-js-compat/src/build-modules-by-versions.js
@@ -1,15 +1,15 @@
 'use strict';
 const { writeFileSync } = require('fs');
 const { resolve } = require('path');
-const data = require('./data');
+const modules = require('../modules');
 const modulesByVersions = require('./modules-by-versions');
 
-const modules = new Set(Object.keys(data));
+const defaults = new Set(modules);
 
 for (const version of Object.values(modulesByVersions)) {
-  for (const module of version) modules.delete(module);
+  for (const module of version) defaults.delete(module);
 }
 
 writeFileSync(resolve(__dirname, '../modules-by-versions.json'), JSON.stringify(Object.assign({
-  '3.0': [...modules],
+  '3.0': [...defaults],
 }, modulesByVersions), null, '  '));