+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## Changelog
+##### Unreleased
+- Generate `core-js-compat/modules` since often we need just the list of `core-js` modules
+
 ##### 3.4.7 - 2019.12.03
 - Fixed an NPM publishing issue